requirements.txt
python-telegram-bot==20.3
pandas==2.0.3
fpdf2==2.7.4
requests==2.31.0
python-dotenv==1.0.0